17.4 lopetus :white_check_mark:

+++ Assets/Scripts/DBScripts/ScriptableObjects.cs
@@ -4,5 +4,5 @@
 public class ScriptableObjects : ScriptableObject
 {
     public string playerName;
-    public int score;
+    public double score;
 }
